refactor(vnext): relocate dedupe + debounce to src/vnext/transport/ (#623 step 2)
Pure relocate, no logic change. Both modules are only used by the Discord
plugin and are transport-layer concerns (per-channel message dedup, rapid
message coalescing). utils/ deferred until a real generic util lands.
